bindings.c: fix size mismatch warnings on casting file info privdata
On 32bit targets, gcc gives lots of warnings like this:

    bindings.c: In function ‘cg_readdir’:
    bindings.c:1752:24: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
      struct file_info *d = (struct file_info *)fi->fh;
                            ^

The problem seems to be that the compiler has to assume we can loose
important bits when converting an 64bit uint into an 32bit pointer.
In our case, its okay, as we know that it could only be filled with
an 32bit pointer, but still those warnings sucks.

Therefore do indirect typecast via uintptr_t, and putting this into
a little helper inline function.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>